Separate gen_struct.rs
